requirements.txt
flask
python-dotenv
fpdf2
aiohttp